pyproject.toml
[tool.poetry]
name = "payment-api"
version = "0.1.0"
description = "Payment API created for our Tech Challenge project"
authors = ["MatheusBLopes <matheusblopes10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
prettyconf = "^2.2.1"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.1"
commitizen = "^3.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
